Address comments #4
