+++ flink-connector-jdbc/src/main/java/org/apache/flink/connector/jdbc/dialect/psql/PostgresDialect.java
@@ -18,114 +18,27 @@
 
 package org.apache.flink.connector.jdbc.dialect.psql;
 
-import org.apache.flink.connector.jdbc.converter.JdbcRowConverter;
-import org.apache.flink.connector.jdbc.dialect.AbstractDialect;
+import org.apache.flink.connector.jdbc.dialect.AbstractPostgresCompatibleDialect;
 import org.apache.flink.connector.jdbc.internal.converter.PostgresRowConverter;
-import org.apache.flink.table.types.logical.LogicalTypeRoot;
 import org.apache.flink.table.types.logical.RowType;
 
-import java.util.Arrays;
-import java.util.EnumSet;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 /** JDBC dialect for PostgreSQL. */
-public class PostgresDialect extends AbstractDialect {
+public class PostgresDialect extends AbstractPostgresCompatibleDialect<PostgresRowConverter> {
 
     private static final long serialVersionUID = 1L;
 
-    // Define MAX/MIN precision of TIMESTAMP type according to PostgreSQL docs:
-    // https://www.postgresql.org/docs/12/datatype-datetime.html
-    private static final int MAX_TIMESTAMP_PRECISION = 6;
-    private static final int MIN_TIMESTAMP_PRECISION = 1;
-
-    // Define MAX/MIN precision of DECIMAL type according to PostgreSQL docs:
-    // https://www.postgresql.org/docs/12/datatype-numeric.html#DATATYPE-NUMERIC-DECIMAL
-    private static final int MAX_DECIMAL_PRECISION = 1000;
-    private static final int MIN_DECIMAL_PRECISION = 1;
-
     @Override
-    public JdbcRowConverter getRowConverter(RowType rowType) {
+    protected PostgresRowConverter rowConverter(RowType rowType) {
         return new PostgresRowConverter(rowType);
     }
 
     @Override
-    public String getLimitClause(long limit) {
-        return "LIMIT " + limit;
+    protected String defaultDriver() {
+        return "org.postgresql.Driver";
     }
 
     @Override
-    public Optional<String> defaultDriverName() {
-        return Optional.of("org.postgresql.Driver");
-    }
-
-    /** Postgres upsert query. It use ON CONFLICT ... DO UPDATE SET.. to replace into Postgres. */
-    @Override
-    public Optional<String> getUpsertStatement(
-            String tableName, String[] fieldNames, String[] uniqueKeyFields) {
-        String uniqueColumns =
-                Arrays.stream(uniqueKeyFields)
-                        .map(this::quoteIdentifier)
-                        .collect(Collectors.joining(", "));
-        String updateClause =
-                Arrays.stream(fieldNames)
-                        .map(f -> quoteIdentifier(f) + "=EXCLUDED." + quoteIdentifier(f))
-                        .collect(Collectors.joining(", "));
-        return Optional.of(
-                getInsertIntoStatement(tableName, fieldNames)
-                        + " ON CONFLICT ("
-                        + uniqueColumns
-                        + ")"
-                        + " DO UPDATE SET "
-                        + updateClause);
-    }
-
-    @Override
-    public String quoteIdentifier(String identifier) {
-        return identifier;
-    }
-
-    @Override
-    public String dialectName() {
+    protected String dialect() {
         return "PostgreSQL";
-    }
-
-    @Override
-    public Optional<Range> decimalPrecisionRange() {
-        return Optional.of(Range.of(MIN_DECIMAL_PRECISION, MAX_DECIMAL_PRECISION));
-    }
-
-    @Override
-    public Optional<Range> timestampPrecisionRange() {
-        return Optional.of(Range.of(MIN_TIMESTAMP_PRECISION, MAX_TIMESTAMP_PRECISION));
-    }
-
-    @Override
-    public Set<LogicalTypeRoot> supportedTypes() {
-        // The data types used in PostgreSQL are list at:
-        // https://www.postgresql.org/docs/12/datatype.html
-
-        // TODO: We can't convert BINARY data type to
-        //  PrimitiveArrayTypeInfo.BYTE_PRIMITIVE_ARRAY_TYPE_INFO in
-        // LegacyTypeInfoDataTypeConverter.
-
-        return EnumSet.of(
-                LogicalTypeRoot.CHAR,
-                LogicalTypeRoot.VARCHAR,
-                LogicalTypeRoot.BOOLEAN,
-                LogicalTypeRoot.VARBINARY,
-                LogicalTypeRoot.DECIMAL,
-                LogicalTypeRoot.TINYINT,
-                LogicalTypeRoot.SMALLINT,
-                LogicalTypeRoot.INTEGER,
-                LogicalTypeRoot.BIGINT,
-                LogicalTypeRoot.FLOAT,
-                LogicalTypeRoot.DOUBLE,
-                LogicalTypeRoot.DATE,
-                LogicalTypeRoot.TIME_WITHOUT_TIME_ZONE,
-                LogicalTypeRoot.TIMESTAMP_WITHOUT_TIME_ZONE,
-                LogicalTypeRoot.TIMESTAMP_WITH_LOCAL_TIME_ZONE,
-                LogicalTypeRoot.ARRAY);
     }
 }
