+++ flink-connector-jdbc/src/main/java/org/apache/flink/connector/jdbc/internal/converter/PostgresRowConverter.java
@@ -31,7 +31,7 @@ public class PostgresRowConverter extends AbstractPostgresCompatibleRowConverter
     private static final long serialVersionUID = 1L;
 
     @Override
-    public String converterName() {
+    protected String name() {
         return "PostgreSQL";
     }
